+++ src/test/java/org/logstash/outputs/s3/GzipRecoverTest.java
@@ -1,4 +1,0 @@
-package test.java.org.logstash.outputs.s3;
-
-public class GzipRecoverTest {
-}